Change the locally defined Db's to use aliases from the ipmiComm db's
Since the defined Db's are just copies of ipmiComm db's, with the name
changed to fit the LNLS name convention, we can replace the record
definition with aliases to the support module record. Also add the macro
`$(sensinst)` in `sensor_ai_alias` that fixes a bug in the previous
implementation where records with the same name would be generated and
one would overwrite the other.