Remove sensinst from PV name.
We were using `sensinst` to avoid duplicated PV names, but it makes a
big mess on sensors that have its indexing on its `attr`. So, it's
better to solve the duplicated names on the `attr` fields, since
`sensinst` is used in the driver code to identify which sensor to read,
and not all sensor numbering scheme will match both.
Duplicate names happen only with our aliases, as ipmiComm base records
themselves also use sensinst in their name. Given that EPICS base does
not allow duplicated aliases by design and complains about it, no extra
setting is required to detect and avoid such duplication issues in the
future.