Redirect power module reading to specific device power monitor PVs
Since the device itself can't read the power consumption, this
information has to come from the power module specific channel that is
related to each device. The power module used is PM2 because PM4 is just
used for redundancy and to replace the PM2 in case of some failure, and
doesn't have any device connected to its channels.