Add SoftReset option to fru_common_alias.
There are no native reset IPMI commands for NAT MCH FRUs, so, in order
to achieve this, we can perform a Power Off and Power On sequence, with
a delay between those two. Furthermore, also is necessary to track the
status of the reset process to ensure that the command will not be
triggered while it's already processing. As the MCH doesn't provide an
IPMI interface to track the FRU MState, it's necessary to track the
command PV itself.
Using the `seq` epics-base record isn't possible here because it not
allow you to be sure the first processing has ended before writing the
second value. For this reason the `sseq` record from `calc` is used.